### 2024-04-25 2a09dbb
updates mga tutorial
