### 2024-06-21 #6655 @prashjha
Howdy! @Jgoldfar, @sevberg, @victoraalves, @fredshone, and @jtock, would you be interested in reviewing the JOSS submission above ([pdf link](https://github.com/openjournals/joss-papers/blob/joss.06655/joss.06655/10.21105.joss.06655.pdf))? The corresponding software is at this [link](https://github.com/arfc/osier).

This [link](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) covers almost everything about the review process; I will be happy to help with more information.

Please let me know if you are interested. Thank you!
