### 2024-06-22 #6655 @fredshone
I can review this

On Fri, 21 Jun 2024 at 19:29, Prashant K. Jha ***@***.***>
wrote:

> Howdy! @jgoldfar <https://github.com/jgoldfar>, @sevberg
> <https://github.com/sevberg>, @victoraalves
> <https://github.com/victoraalves>, @fredshone
> <https://github.com/fredshone>, and @jtock <https://github.com/jtock>,
> would you be interested in reviewing the JOSS submission above (pdf link
> <https://github.com/openjournals/joss-papers/blob/joss.06655/joss.06655/10.21105.joss.06655.pdf>)?
> The corresponding software is at this link <https://github.com/arfc/osier>
> .
>
> This link <https://joss.readthedocs.io/en/latest/reviewer_guidelines.html>
> covers almost everything about the review process; I will be happy to help
> with more information.
>
> Please let me know if you are interested. Thank you!
>
