### 2024-06-22 #6919 @editorialbot
[REVIEW]: Osier: A Python package for multi-objective energy system optimization
**Submitting author:** @samgdotson (<a href="http://orcid.org/0000-0002-8662-0336">Samuel Dotson</a>)
**Repository:** https://github.com/arfc/osier
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.3.1
