### 2024-06-25 arfc/osier#55 @fredshone
Failed pip install
Failed pip install using latest python.

Details:

Windows (10)
`python -m pip install osier pyomo==6.4.1`

I tried:

Fixed by back-tracking to 3.10.

BUT please clarify expected supported python versions in installation instructions.

Suggest you also explicitly add version of python to CI.
