### 2024-06-25 arfc/osier#56 @fredshone
Failed tests on Windows (10)
Following "developer" installation I had failing tests due to missing solver:

```
git clone git@github.com:arfc/osier.git
cd osier
pip install .
pip install .[doc]
pip install pyomo==6.4.1
pytest
```

Results in (for example) `pyomo.common.errors.ApplicationError: No executable found for solver 'cplex'`.

Please move solver installation instructions from "Getting Started" to "Installation".

To allow a smoother onboarding for windows users I suggest adding installation instructions for glpk (eg `conda install -c conda-forge glpk`). I appreciate this might not be the best solver but it is more easilly accessible which is nice for casual users and will be easier for windows based CI.
