### 2024-06-25 arfc/osier#57 @fredshone
Example accessibility
Nice example notebooks! i found them in the docs and wanted to have a play with them. I encountered two blockers:

- was expecting to find them in an examples directory near root (not in docs)
- `jupyter` is not included in the requirements
- required kernel installation (`ipython kernel install --user --name=osier`)

Note that the latter two (especially the last) are quite specific to people using an env/package manager.
