### 2024-06-25 arfc/osier#58 @fredshone
No env control in installation instructions
I think it is a good idea to include an env/package manager such as conda or [mamba ](https://mamba.readthedocs.io/en/latest/index.html)in the installation instructions.

- This will create more homogenous experience for users on different OS
- You should get less awkward bug reports from less expereinced users

There is an example of a (also energy) project that makes use of mamba [here](https://github.com/calliope-project/calliope).
