### 2024-06-25 #6919 @fredshone
Really nice work well presented.

I require some changes detailed below with optional actions mixed in. A lot of these are based on compatability with Windows. Although not required I suggest both adding Windows to your CI and including env control in the docs. [Mamba](https://mamba.readthedocs.io/en/latest/index.html) is very good. There are example projects [here](https://github.com/arup-group/pam) and [here](https://github.com/calliope-project/calliope).
_______________________________________

**Installation** and **Installation Instructions:**

On a Windows machine I hit two **blockers**; (i) python version (issue [here](https://github.com/arfc/osier/issues/55)), and (ii) solver installation documentation (issue [here](https://github.com/arfc/osier/issues/56)).

I strongly recommend adding a windows machine to the CI. Otherwise this is likely to reoccur.

I have also suggested that the default Windows solver be glpk. This is easy to install and will smooth CI. It is also more in keeping with JOSS (in my opinion). if you stick with a less open solver I think this needs to be explicitly mentioned in ther paper.

OPTIONAL, I also encourage use of mamba in the install instructions as per [this ](https://github.com/arfc/osier/issues/58)issue.
__________________

**Example usage:** OPTIONAL - there is only a single example of multi-criteria and it is a little hidden at the end of a notebook. Please consider making this more obvious. At the moment multi-criteria appears as an after-thought rather than a core feature. OPTIONAL - please consider making the example notebooks more accessible in the project rather than just docs. See [this](https://github.com/arfc/osier/issues/57) issue.
______________________

**Automated tests:** Please add a code coverage report (eg pytest-cov).
_________________________

**State of the field** and **References**: Please provide some more specific description and reference to similar tools. I think your focus on multi-criteria is very important, as you state - there are many other open energy system modelling tools. You would make the case better by more explicit comparison to similar projects that don't do multi-criteria. For example [calliope](https://github.com/calliope-project/calliope) is very similar in many ways but does not to multi-criteria without hacking the solver.
