### 2024-08-29 #6919 @victoraalves
I have performed my initial review.

The tool is interesting and a good contribution! Congratulations, nice work.

I had some issues regarding Python version compatibility (already mentioned  [here](https://github.com/arfc/osier/issues/55) by the other reviewer), as well as problems related to numpy 2.0 that breaks some of the functionality of the tool.

### Installation

Installation on macOS was a bit problematic, due to the new Python versions and the Numpy version installed by default when doing "pip install". This needs to be clarified in the documentation and/or fixed in the project as well.

### Testing

Some tests failed due to the lack of the "cbc" solver, despite the fact that I followed the installation instructions as recommended.

### Status of the field

As mentioned, I believe it's a good idea to highlight other tools that have the same (or similar) objectives and explain how Osier differentiates itself.
