### 2024-09-09 arfc/osier#55 @samgdotson
Thanks for opening this issue @fredshone and confirming the problem @victoraalves. I think this issue is related to #50 and #59. Pyomo appears to be incompatible with NumPy 2.0, at the moment. https://github.com/Pyomo/pyomo/issues/3290
