### 2024-09-10 1446f80
removes setup.py
