### 2024-09-10 744e365
first attempt at pyproject.toml
