+++ pyproject.toml
@@ -1,0 +1,69 @@
+[build-system]
+requires = ["setuptools >= 64.0"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name="osier"
+version = "0.3.0"
+description = "osier: A justice oriented energy system optimization tool"
+readme = "README.md"
+keywords = ["energy systems", "optimization", "multi-objective", "justice", "multi-criteria decision"]
+
+license = { file = "LICENSE" }
+
+requires-python = ">= 3.9"
+
+dependencies = [
+    'numpy',
+    'pandas',
+    'matplotlib',
+    'pytest',
+    'dill',
+    'openpyxl',
+    'nrelpy',
+    'unyt',
+    'pymoo',
+    'pyentrp',
+    'deap',
+    'pyomo'
+]
+
+authors = [
+  {name = "Sam Dotson", email = "samgdotson@gmail.com"},
+]
+maintainers = [
+  {name = "Sam Dotson", email = "samgdotson@gmail.com"}
+]
+
+classifiers = [
+    "Development Status :: 3 - Alpha",
+    "Intended Audience :: Science/Research",
+    "Programming Language :: Python :: 3.9",
+    "Programming Language :: Python :: 3.10",
+    "Programming Language :: Python :: 3.11",
+    "Programming Language :: Python :: 3.12",
+    "Environment :: Console",
+    "Natural Language :: English",
+    "Operating System :: OS Independent"
+]
+
+[project.urls]
+Homepage = "https://osier.readthedocs.io"
+Repository = "https://github.com/arfc/osier"
+Issues = "https://github.com/arfc/osier/issues"
+
+[project.optional-dependencies]
+docs = [
+        'sphinx>=5.1',
+        'sphinx-autobuild',
+        'myst-parser',
+        "sphinx_design",
+        "sphinx-autodoc-typehints",
+        'numpydoc',
+        'pydata_sphinx_theme',
+        'nbsphinx',
+        'pandoc'
+]
+
+[tool.setuptools.packages.find]
+include = ['osier']
