### 2024-09-10 2eb861d
set the log level for the root logger
