+++ README.md
@@ -12,7 +12,7 @@ Open source multi-objective energy system framework
 
 `osier` is available through [PyPI](https://pypi.org/project/osier/). It may be installed with 
 ```bash
-python -m pip install osier pyomo==6.4.1
+python -m pip install osier
 ``` 
 or by cloning this repository and building from source:
 
@@ -25,18 +25,8 @@ cd osier
 pip install .
 # to also install the documentation dependencies
 pip install .[doc]
-
-# followed by 
-pip install pyomo==6.4.1
 ```
 
-
-```{note}
-Although `pyomo` is a dependency, the current version of `pyomo` (6.7.1, as of 2/29/24) has a bug
-that prints erroneous errors during an `osier` simulation. Therefore, users are recommended to 
-install a specific version of `pyomo` after the main installation of `osier`. There is an open issue [#50](https://github.com/arfc/osier/issues/50) 
-related to this concern.
-```
 
 ## Documentation
 The documentation for `osier` can be viewed [here](https://osier.readthedocs.io/en/latest/). 
