### 2024-09-10 8f4e0c5
adds new workflows to ci
