+++ .github/workflows/CI.yml
@@ -15,7 +15,15 @@ permissions:
 jobs:
   build:
     name: Build and test
-    runs-on: ubuntu-latest
+    runs-on: ${{ matrix.os }}
+    strategy:
+      fail-fast: false
+      matrix:
+        python-version: ${{ fromJSON(needs.build.outputs.python-versions) }}
+        os:
+        - ubuntu-latest
+        - macos-latest
+        - windows-latest
 
     defaults:
       run:
@@ -36,11 +44,10 @@ jobs:
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
-        pip install flake8 pytest pyomo==6.4.1
         if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
     - name: Install CBC
       run: |
-        mamba install coincbc
+        mamba install coin-or-cbc coincbc
     - name: Lint with flake8
       run: |
         # stop the build if there are Python syntax errors or undefined names
@@ -50,9 +57,6 @@ jobs:
     - name: Install osier package
       run: |
         pip install .
-    - name : Install recent unyt update
-      run: |
-        python3 -m pip install -e git+https://github.com/yt-project/unyt.git#egg=unyt
     - name: Test with pytest
       run: |
         pytest tests/
