### 2024-09-10 7a10343
updates the ci workflow
