+++ pyproject.toml
@@ -53,7 +53,7 @@ Repository = "https://github.com/arfc/osier"
 Issues = "https://github.com/arfc/osier/issues"
 
 [project.optional-dependencies]
-docs = [
+doc = [
         'sphinx>=5.1',
         'sphinx-autobuild',
         'myst-parser',
