+++ .github/workflows/CI.yml
@@ -20,10 +20,10 @@ jobs:
       fail-fast: false
       matrix:
         python-version: 
-        - 3.9
-        - 3.10
-        - 3.11
-        - 3.12
+        - "3.9"
+        - "3.10"
+        - "3.11"
+        - "3.12"
         os:
         - ubuntu-latest
         - macos-latest
