### 2024-09-10 4ec3c14
adds flake8 installation
