+++ .github/workflows/CI.yml
@@ -49,6 +49,9 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
+    - name: Install flake8
+      run: |
+        python -m pip install flake8
     - name: Install CBC
       if: matrix.os != 'windows-latest'
       run: |
@@ -59,12 +62,6 @@ jobs:
         curl -L https://github.com/coin-or/Cbc/releases/download/releases%2F2.10.10/Cbc-releases.2.10.10-w64-msvc17-md.zip -o cbc.zip
         unzip cbc.zip -d ${HOME}/cbc
         echo "${HOME}/cbc/bin" >> $GITHUB_PATH
-    - name: Lint with flake8  
-      run: |
-        # stop the build if there are Python syntax errors or undefined names
-        flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
-        # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
-        flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
     - name: Install osier package (macos)
       if: matrix.os == 'macos-latest'
       run: |
@@ -77,3 +74,10 @@ jobs:
     - name: Test with pytest
       run: |
         pytest tests/
+
+    - name: Lint with flake8  
+      run: |
+        # stop the build if there are Python syntax errors or undefined names
+        flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
+        # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
+        flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
