### 2024-09-10 e2ecd30
updates the default solver to cbc
