+++ .github/workflows/CI.yml
@@ -51,7 +51,7 @@ jobs:
         if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
     - name: Install flake8
       run: |
-        python -m pip install flake8
+        python -m pip install flake8  
     - name: Install CBC
       if: matrix.os != 'windows-latest'
       run: |
@@ -66,7 +66,7 @@ jobs:
       if: matrix.os == 'macos-latest'
       run: |
         pip install -e .'[doc]'
-    - name: Install osier package (macos)
+    - name: Install osier package
       if: matrix.os != 'macos-latest'
       run: |
         pip install -e .[doc]
