### 2024-09-10 ccb308d
changes windows solver to cbc to make CI work
