+++ tests/test_equations.py
@@ -7,7 +7,7 @@ import sys
 import functools
 
 if "win32" in sys.platform:
-    solver = 'cplex'
+    solver = 'cbc'
 elif "linux" in sys.platform:
     solver = "cbc"
 else:
