### 2024-09-10 e31d6ab
adds two articles to bib and reinstates the joss draft build step
