+++ paper/paper.bib
@@ -210,3 +210,44 @@
   keywords = {electricity scenarios,energy modelling,Jordan,MCA,MENA,Morocco,multi-criteria,stakeholder participation,sustainability,Tunisia},
   file = {/home/sam/snap/zotero-snap/common/Zotero/storage/S3QTXVF7/Zelt et al. - 2019 - Long-Term Electricity Scenarios for the MENA Regio.pdf}
 }
+
+@article{brown:2018,
+	title = {{PyPSA}: {Python} for {Power} {System} {Analysis}},
+	volume = {6},
+	copyright = {Authors who publish with this journal agree to the following terms:    Authors retain copyright and grant the journal right of first publication with the work simultaneously licensed under a  Creative Commons Attribution License  that allows others to share the work with an acknowledgement of the work's authorship and initial publication in this journal.  Authors are able to enter into separate, additional contractual arrangements for the non-exclusive distribution of the journal's published version of the work (e.g., post it to an institutional repository or publish it in a book), with an acknowledgement of its initial publication in this journal.  Authors are permitted and encouraged to post their work online (e.g., in institutional repositories or on their website) prior to and during the submission process, as it can lead to productive exchanges, as well as earlier and greater citation of published work (See  The Effect of Open Access ).  All third-party images reproduced on this journal are shared under Educational Fair Use. For more information on  Educational Fair Use , please see  this useful checklist prepared by Columbia University Libraries .   All copyright  of third-party content posted here for research purposes belongs to its original owners.  Unless otherwise stated all references to characters and comic art presented on this journal are ©, ® or ™ of their respective owners. No challenge to any owner’s rights is intended or should be inferred.},
+	issn = {2049-9647},
+	shorttitle = {{PyPSA}},
+	url = {http://openresearchsoftware.metajnl.com/articles/10.5334/jors.188/},
+	doi = {10.5334/jors.188},
+	abstract = {Python for Power System Analysis (PyPSA) is a free software toolbox for simulating and optimising modern electrical power systems over multiple periods. PyPSA includes models for conventional generators with unit commitment, variable renewable generation, storage units, coupling to other energy sectors, and mixed alternating and direct current networks. It is designed to be easily extensible and to scale well with large networks and long time series. In this paper the basic functionality of PyPSA is described, including the formulation of the full power flow equations and the multi-period optimisation of operation and investment with linear power flow equations. PyPSA is positioned in the existing free software landscape as a bridge between traditional power flow analysis tools for steady-state analysis and full multi-period energy system models. The functionality is demonstrated on two open datasets of the transmission system in Germany (based on SciGRID) and Europe (based on GridKit).Funding statement: This research was conducted as part of the CoNDyNet project, which is supported by the German Federal Ministry of Education and Research under grant no. 03SF0472C. The responsibility for the contents lies solely with the authors},
+	language = {en},
+	number = {1},
+	urldate = {2022-12-12},
+	journal = {Journal of Open Research Software},
+	author = {Brown, Thomas and Hörsch, Jonas and Schlachtberger, David},
+	month = jan,
+	year = {2018},
+  publisher = Ubiquity Press,
+	keywords = {energy system simulations, Load flow calculations, optimal power flow, Power system simulations, renewable energy, security-constrained optimal power flow, unit commitment},
+	pages = {4},
[…]
