+++ .github/workflows/CI.yml
@@ -81,3 +81,6 @@ jobs:
         flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
         # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
         flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
+
+    - name: Open Journals PDF Generator
+      uses: openjournals/openjournals-draft-action@v.1.0
