### 2024-09-10 615c9dd
adds a line about two related projects, calliope and pypsa
