### 2024-09-14 d91646b
comments out openjournals hook in CI file
