+++ .github/workflows/CI.yml
@@ -82,5 +82,5 @@ jobs:
         # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
         flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
 
-    - name: Open Journals PDF Generator
-      uses: openjournals/openjournals-draft-action@v.1.0
+    # - name: Open Journals PDF Generator
+    #   uses: openjournals/openjournals-draft-action@v.1.0
