### 2024-09-14 3116516
adds pull_request to draft pdf workflow
