+++ paper/paper.bib
@@ -227,7 +227,6 @@
 	author = {Brown, Thomas and Hörsch, Jonas and Schlachtberger, David},
 	month = jan,
 	year = {2018},
-  publisher = Ubiquity Press,
 	keywords = {energy system simulations, Load flow calculations, optimal power flow, Power system simulations, renewable energy, security-constrained optimal power flow, unit commitment},
 	pages = {4},
 	file = {Full Text PDF:/Users/samdotson/Zotero/storage/WXM33X5R/Brown et al. - 2018 - PyPSA Python for Power System Analysis.pdf:application/pdf},
