### 2024-09-14 d7d1297
fixes erroneous 'download' in the 'upload' rule
