### 2024-09-14 9c69524
adds scipy to requirements
