+++ pyproject.toml
@@ -16,6 +16,7 @@ requires-python = ">= 3.9"
 dependencies = [
     'numpy',
     'pandas',
+    'scipy',
     'matplotlib',
     'pytest',
     'dill',
