+++ environment.yml
@@ -1,0 +1,26 @@
+name: osier
+channels:
+  - conda-forge
+
+dependencies:
+  - python<=3.11
+  - pip 
+  - ipython
+  - ipykernel
+  - pandas
+  - scipy
+  - glpk
+  - unyt
+  - matplotlib
+  - yaml
+  - numpy
+  - matplotlib
+  - pytest
+  - pytest-cov
+  - pyomo
+  - pip:
+    - pymoo
+    - pyentrp
+    - deap
+    - dill
+    - openpyxl
