### 2024-09-15 0389bdd
removes maximum python version, reruns example
