+++ environment.yml
@@ -1,12 +1,13 @@
-name: osier
+name: osier-env
 channels:
   - conda-forge
 
 dependencies:
-  - python<=3.11
+  - python>=3.9
   - pip 
   - ipython
   - ipykernel
+  - jupyterlab
   - pandas
   - scipy
   - glpk
