### 2024-09-16 0ae7a21
adds coverage and failunder to pytest command
