+++ pyproject.toml
@@ -18,6 +18,7 @@ dependencies = [
     'pandas',
     'matplotlib',
     'pytest',
+    'pytest-cov',
     'dill',
     'openpyxl',
     'nrelpy',
@@ -67,3 +68,9 @@ doc = [
 
 [tool.setuptools.packages.find]
 include = ['osier']
+
+[tool.pytest.ini_options]
+addopts = "--cov"
+
+[tool.coverage.report]
+fail_under = 80
