### 2024-09-19 6ecbe8f
adds nbsphinx-link
