### 2024-09-23 #6919 @samgdotson
@fredshone @victoraalves

Thank you both for your reviews! @fredshone, thank you for making a pull request, as well.

Regarding the **status of the field**, how much elaboration are you both looking for? The paper is already ~1000 words. I have an [open pull request](https://github.com/arfc/osier/pull/66) that adds a direct reference to two open source packages. Is more elaboration necessary will this suffice?

> Two well known open-source ESOMs, Calliope [@pfenninger:2018] and Python for Power Systems Analysis (PyPSA)
> [@brown:2018], partially address equity issues by implementing MGA, but this does not resolve the limitation of
> mono-objective optimization.

Please let me know!
