### 2024-09-23 #6919 @fredshone
@samgdotson short is good I agree. (from memory) I'd like to see specific mention on how your work builds on above by adding adding multi-objective functionality. This could be a one-liner.

FYI, I spoke to calliope authors and they agreed MO was not a proper or intended functionality. So lean on it hard.

More pressing for me personally is that when I went through the examples - the (single) multi-criteria example was at the end of a long notebook. I'd like to see it (perhaps with an even more minimal example) brought more front and center so that a new user (me in this case) can more quickly use the examples to understand the project scope.
