### 2024-09-23 1e43c91
Apply suggestions from code review
