### 2024-09-29 6ea8618
adds new notebook and osier to environment file
