+++ environment.yml
@@ -11,6 +11,7 @@ dependencies:
   - pandas
   - scipy
   - glpk
+  - coincbc
   - unyt
   - matplotlib
   - yaml
@@ -25,3 +26,4 @@ dependencies:
     - deap
     - dill
     - openpyxl
+    - git+https://github.com/arfc/osier.git
