### 2024-10-04 01b7caf
updates env file and new notebook
