### 2024-10-04 0783f43
updates environment.yml file
