### 2024-10-04 7ca2432
adds conda env instructions and binder badge
