### 2024-10-04 506f02e
adds examples instruction to the readme
