### 2024-10-08 #6919 @samgdotson
@fredshone @victoraalves I pushed a [final PR](https://github.com/arfc/osier/pull/70) that should address the remaining comments.

This PR
* adds a notebook with a "minimum working example" (which I suppose is only "minimal" in the energy modeling space 😄) called "getting started."
* adds an `environment.yml` file and instructions to use this file to install `osier` -- this should cover the env control request from @fredshone
* addresses the notebook access by adding a binder badge to each example notebook in the repository (I couldn't get nblink to work and I don't want two version of the examples floating around, perhaps this can be added in the future).

If you both agree and are satisfied, what are the next steps? Thanks for your reviews!
