### 2024-11-25 #6919 @samgdotson
Hi @prashjha, here is the Zenodo publication is available here: 10.5281/zenodo.14216170

[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.14216170.svg)](https://doi.org/10.5281/zenodo.14216170)

The current version of the code is v0.3.1, thanks!

Let me know if I need to do anything else.
