### 2024-11-25 #6919 @editorialbot
:wave: @openjournals/pe-eics, this paper is ready to be accepted and published.
