### 2024-12-04 #6919 @samgdotson
Hi @prashjha! Is there anything else I need to do at this moment to finalize the acceptance? Thanks!
