### 2024-12-05 #6919 @kyleniemeyer
@samgdotson just a few things in the paper to fix:
- a space is missing between "dataset" and the NREL citation
- the 2023 NREL citation needs some more information, like a DOI or URL where the report can be found
