+++ paper/paper.bib
@@ -119,11 +119,12 @@
   file = {/home/sam/snap/zotero-snap/common/Zotero/storage/N4X7U5L5/McKenna et al. - 2018 - Combining local preferences with multi-criteria de.pdf;/home/sam/snap/zotero-snap/common/Zotero/storage/DFTWLY3S/S0377221718300729.html}
 }
 
-@misc{nationalrenewableenergylaboratory:2023,
+@dataset{nationalrenewableenergylaboratory:2023,
   title = {2023 {{Annual Technology Baseline}} ({{ATB}})},
   author = {{National Renewable Energy Laboratory}},
   year = {2023},
   urldate = {2024-02-26},
+  url={https://atb.nrel.gov/}
   file = {/home/sam/snap/zotero-snap/common/Zotero/storage/URRFWYVN/data.html}
 }
 
