### 2024-12-05 8390636
adds missing comma to bib file
