+++ paper/paper.bib
@@ -124,7 +124,7 @@
   author = {{National Renewable Energy Laboratory}},
   year = {2023},
   urldate = {2024-02-26},
-  url={https://atb.nrel.gov/}
+  url={https://atb.nrel.gov/},
   file = {/home/sam/snap/zotero-snap/common/Zotero/storage/URRFWYVN/data.html}
 }
 
