+++ paper/paper.bib
@@ -119,7 +119,7 @@
   file = {/home/sam/snap/zotero-snap/common/Zotero/storage/N4X7U5L5/McKenna et al. - 2018 - Combining local preferences with multi-criteria de.pdf;/home/sam/snap/zotero-snap/common/Zotero/storage/DFTWLY3S/S0377221718300729.html}
 }
 
-@dataset{nationalrenewableenergylaboratory:2023,
+@misc{nationalrenewableenergylaboratory:2023,
   title = {2023 {{Annual Technology Baseline}} ({{ATB}})},
   author = {{National Renewable Energy Laboratory}},
   year = {2023},
