### 2024-12-05 fe6d9b6
adds place to nrel ref
