+++ paper/paper.bib
@@ -125,6 +125,7 @@
   year = {2023},
   urldate = {2024-02-26},
   url={https://atb.nrel.gov/},
+  place={Golden, CO: National Renewable Energy Laboratory}
   file = {/home/sam/snap/zotero-snap/common/Zotero/storage/URRFWYVN/data.html}
 }
 
