### 2024-12-05 068d746
adds missing comma to bib file
