### 2024-12-06 #6919 @samgdotson
Thanks @kyleniemeyer! I updated the bib file and the reference is now consistent with the recommended citation found [here](https://atb.nrel.gov/electricity/2023/about).

> Recommended Citation
> We recommend you cite the 2023 ATB as NREL (2023) and reference it as:
>
> NREL (National Renewable Energy Laboratory). 2023. "2023 Annual Technology Baseline." Golden, CO: National
> Renewable Energy Laboratory. https://atb.nrel.gov/.
