### 2024-12-06 #6919 @kyleniemeyer
@samgdotson ah, I missed this until after it was published: can you remove the `<code>` fields around the software name in the paper's title? Our system cannot process that.
